Fixed wildcard matcher
